requirements.txt
Pillow>=10.0.0
pymupdf>=1.24.0
